requirements.txt
Babel==2.10.1
bcrypt==3.2.2
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-HTTPAuth==4.6.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
Pillow==9.1.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
speaklater==1.3
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
wtf==0.1
WTForms==3.0.1
